+++ src/main/java/uk.gov.hmcts.reform.fpl.ccddatamigration/CaseMigrationProcessor.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.service;
+package uk.gov.hmcts.reform.fpl.ccddatamigration;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.fpl.ccddatamigration.ccd.CoreCaseDataService;
+import uk.gov.hmcts.reform.fpl.ccddatamigration.service.DataMigrationService;
 
 import java.util.ArrayList;
 import java.util.List;
