+++ src/test/java/uk/gov/hmcts/reform/fpl/ccddatamigration/CaseMigrationProcessorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.service;
+package uk.gov.hmcts.reform.fpl.ccddatamigration;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -9,6 +9,7 @@ import org.springframework.util.ReflectionUtils;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.PaginatedSearchMetadata;
+import uk.gov.hmcts.reform.fpl.ccddatamigration.CaseMigrationProcessor;
 import uk.gov.hmcts.reform.fpl.ccddatamigration.ccd.CoreCaseDataService;
 
 import java.lang.reflect.Field;
